add more feature

+++ pages/user/index.js
@@ -116,14 +116,15 @@ export default function user() {
                                         </td>
                                         <td className="py-3 px-6 text-center">
                                             <div className="flex item-center justify-center">
-                                                <div
+                                                <a
+                                                    href="#edit"
                                                     onClick={() =>
                                                         router.push(
                                                             "/user/edit/" +
                                                                 i._id
                                                         )
                                                     }
-                                                    className="w-4 mr-2 transform hover:text-primary hover:scale-110"
+                                                    className="w-4 mr-2 transform hover:text-primary hover:scale-125"
                                                 >
                                                     <svg
                                                         xmlns="http://www.w3.org/2000/svg"
@@ -138,14 +139,24 @@ export default function user() {
                                                             d="M15.232 5.232l3.536 3.536m-2.036-5.036a2.5 2.5 0 113.536 3.536L6.5 21.036H3v-3.572L16.732 3.732z"
                                                         />
                                                     </svg>
-                                                </div>
-                                                <div
-                                                    className="w-4 mr-2 transform hover:text-primary hover:scale-110"
-                                                    onClick={() =>
-                                                        deleteData(i._id)
-                                                    }
+                                                </a>
+                                                <a
+                                                    href="#delete"
+                                                    className="w-4 mr-2 transform hover:text-primary hover:scale-125"
+                                                    onClick={() => {
+                                                        if (
+                                                            window.confirm(
+                                                                "Apakah anda yakin akan mengapus karyawan " +
+                                                                    i.name +
+                                                                    "?"
+                                                            )
+                                                        ) {
+                                                            deleteData(i._id)
+                                                        }
+                                                    }}
                                                 >
                                                     <svg
+                                                        className="text-red-500"
                                                         xmlns="http://www.w3.org/2000/svg"
                                                         fill="none"
                                                         viewBox="0 0 24 24"
@@ -158,7 +169,7 @@ export default function user() {
                                                             d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"
                                                         />
                                                     </svg>
-                                                </div>
+                                                </a>
                                             </div>
                                         </td>
                                     </tr>
